Split terminal file link detection modules
